requirements.txt
langchain 
wikipedia 
langchain-community 
langchain-groq
python-dotenv
pydantic
duckduckgo-search